Moved model tests to the model package
