+++ richtextfx/src/test/java/org/fxmisc/richtext/model/ParagraphTest.java
@@ -1,8 +1,6 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import static org.junit.Assert.*;
-
-import org.fxmisc.richtext.model.Paragraph;
 import org.junit.Test;
 
 public class ParagraphTest {
