+++ richtextfx/src/test/java/org/fxmisc/richtext/model/ReadOnlyStyledDocumentTest.java
@@ -1,12 +1,9 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import static org.fxmisc.richtext.model.ReadOnlyStyledDocument.*;
 import static org.junit.Assert.*;
 
 import java.util.List;
-
-import org.fxmisc.richtext.model.Paragraph;
-import org.fxmisc.richtext.model.ReadOnlyStyledDocument;
 import org.junit.Test;
 
 public class ReadOnlyStyledDocumentTest {
