+++ richtextfx/src/test/java/org/fxmisc/richtext/model/StyledTextAreaModelTest.java
@@ -1,12 +1,10 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import static org.junit.Assert.*;
 
 import java.util.Collection;
 import java.util.Collections;
 
-import org.fxmisc.richtext.model.SimpleEditableStyledDocument;
-import org.fxmisc.richtext.model.StyledTextAreaModel;
 import org.junit.Test;
 
 public class StyledTextAreaModelTest {
